+++ awto-log-ws/src/main/java/cl/awto/hastag/vo/ResponseId.java
@@ -1,8 +1,0 @@
-package cl.awto.hastag.vo;
-
-import java.io.Serializable;
-
-@Deprecated
-public class ResponseId implements Serializable {
-    private Integer id;
-}